refactor(agent): extract Copilot credential error classifier
